feature: implement GetFileInfo, listDirectoryChildren with input validation and unit test
feature: add inline exception handling in controller to ensure JSON-RPC error responses include request id, refer to error code standard in JSON RPC 2.0 doc.

+++ src/test/java/com/jetbrains/filesystem/FileManageServiceTest.java
@@ -44,9 +44,15 @@ public class FileManageServiceTest {
     void testListDirectoryChildren() throws Exception {
         String testDir = "folder02";
         Path dirPath = Paths.get(fileServiceProperties.getRootFolder(),testDir);
-        Files.createDirectories(dirPath);
-        Files.writeString(dirPath.resolve("file02.txt"), "Hello World");
-        Files.createDirectory(dirPath.resolve("subfolder02"));
+        if (!Files.exists(dirPath)) {
+            Files.createDirectories(dirPath);
+        }
+        if (!Files.exists(dirPath.resolve("file02.txt"))) {
+            Files.writeString(dirPath.resolve("file02.txt"), "Hello World");
+        }
+        if (!Files.exists(dirPath.resolve("subfolder02"))) {
+            Files.createDirectory(dirPath.resolve("subfolder02"));
+        }
 
         List<FileInfo> fileInfoList = fileManageService.listDirectoryChildren(testDir);
 
@@ -63,6 +69,8 @@ public class FileManageServiceTest {
         }
         assertTrue(hasFile);
         assertTrue(hasSubfolder);
+
+
     }
 
 }
